[A.0.1] Works-ish
Due to limited acces to data, this is as complete as it will be unless I can find another way to access individual scores (particularly those without written reviews)
Currently the program reads in the data from the website's html code for the pages with user reviews.
Need to look into getting permission to use the official API.. will that even have what I need?